QWizard: Replace another pointer table with char arrays
Like in I53284066, the result of the SIGNALS() macro is handled in a switch
statement, while the other two character pointers in the struct are replaced
by character arrays of 'maximum occurring size'. If this looks wasteful, it
really isn't:

 Linux AMD64 GCC 4.9-pre stripped -O2 effects:
     text size: -280B
     data size: -160B
     relocs:     -21

When adding longer strings, compilers will warn, so this doesn't hurt
maintainability, either.

Change-Id: I5ac1cdffd8ac0ea0a1ede1ea4edcc6d3e22dcaa2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>